// angular-electron/src/app/models/DefaultResponse.ts
/*
 *    Copyright 2023, Sergio Lissner
 *
 *    Licensed under the Apache License, Version 2.0 (the "License");
 *    you may not use this file except in compliance with the License.
 *    You may obtain a copy of the License at
 *
 *        http://www.apache.org/licenses/LICENSE-2.0
 *
 *    Unless required by applicable law or agreed to in writing, software
 *    distributed under the License is distributed on an "AS IS" BASIS,
 *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *    See the License for the specific language governing permissions and
 *    limitations under the License.
 */

import { OperationStatus } from '@app/enums/OperationStatus';

export interface DefaultResponse {
    errorMessages: string[];
    infoMessages: string[];
    errorMessagesAsStr: string;
    errorMessagesAsList: string[];
    infoMessagesAsList: string[];

    status?: OperationStatus;
}